+++ app/Http/Middleware/EnsureNotInstalled.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace App\Http\Middleware;
+
+use Closure;
+use Illuminate\Http\Request;
+use Symfony\Component\HttpFoundation\Response;
+
+class EnsureNotInstalled
+{
+    public function handle(Request $request, Closure $next): Response
+    {
+        if (file_exists(base_path('.installed'))) {
+            abort(404);
+        }
+
+        return $next($request);
+    }
+}
+
+